@@ -1,11 +1,11 @@
 name: app
 description: "J’agis vous accompagne à chaque étape de la transition écologique."
 publish_to: 'none'
-version: 0.16.21+0
+version: 0.16.22+0
 
 environment:
-  sdk: 3.7.0
-  flutter: 3.29.1
+  sdk: 3.7.2
+  flutter: 3.29.2
 
 dependencies:
   clock: 1.1.2
@@ -41,8 +41,8 @@ dependencies:
   package_info_plus: 8.3.0
   pinput: 5.0.1
   rxdart: 0.28.0
-  sentry_dio: 8.13.3
-  sentry_flutter: 8.13.3
+  sentry_dio: 8.14.0
+  sentry_flutter: 8.14.0
   url_launcher: 6.3.1
 
 dev_dependencies:
