@@ -1,4 +1,4 @@
 import 'package:flutter/widgets.dart';
 
-double screenWidth(final BuildContext context, final double? pct) =>
-    pct == null ? MediaQuery.of(context).size.width : pct * MediaQuery.of(context).size.width;
+/// Returns the screen width based on the given percentage of the screen width.
+double screenWidth(final BuildContext context, {final double percentage = 1}) => percentage * MediaQuery.sizeOf(context).width;
