@@ -44,7 +44,7 @@ class _FnvLoaderState extends State<FnvLoader> with SingleTickerProviderStateMix
   }
 
   @override
-  Widget build(final BuildContext context) {
+  Widget build(final context) {
     const amplitude = 75.0;
     const ballWidth = 40.0;
 
