@@ -26,7 +26,8 @@ class ActionRepository {
     return switch (type) {
       ActionType.simulator => Right(ActionSimulatorMapper.fromJson(json)),
       ActionType.classic => Right(ActionClassicMapper.fromJson(json)),
-      ActionType.quiz || ActionType.performance =>
+      ActionType.quiz => Right(ActionQuizMapper.fromJson(json)),
+      ActionType.performance =>
         // TODO(erolley): Handle this case.
         throw UnimplementedError(),
     };
@@ -34,7 +35,7 @@ class ActionRepository {
 
   Future<void> markAsDone({required final ActionType type, required final String id}) async {
     final actionTypeAPI = actionTypeToAPIString(type);
-    // NOTE(erolley): we may want to handle the response here.
+    // TODO(erolley): we may want to handle the response here.
     await _client.post(Endpoints.actionFaite(type: actionTypeAPI, code: id));
   }
 }
