@@ -1,12 +1,12 @@
 import 'package:equatable/equatable.dart';
 
 class Partner extends Equatable {
-  const Partner({required this.nom, required this.url, required this.logo});
+  const Partner({required this.name, required this.logo, required this.url});
 
-  final String nom;
-  final String? url;
+  final String name;
   final String logo;
+  final String? url;
 
   @override
-  List<Object?> get props => [nom, url, logo];
+  List<Object?> get props => [name, logo, url];
 }
