@@ -22,7 +22,7 @@ class _View extends StatelessWidget {
   const _View();
 
   @override
-  Widget build(final BuildContext context) => BlocBuilder<CarSimulatorResultBloc, CarSimulatorResultState>(
+  Widget build(final context) => BlocBuilder<CarSimulatorResultBloc, CarSimulatorResultState>(
     builder:
         (final context, final state) => switch (state) {
           CarSimulatorInit() => const Center(
@@ -61,7 +61,7 @@ class _CarSimulatorResultView extends StatelessWidget {
   final CarSimulatorOption? bestEmissionsOption;
 
   @override
-  Widget build(final BuildContext context) =>
+  Widget build(final context) =>
   // TODO(erolley): ListView doesn't work here
   Padding(
     padding: const EdgeInsets.symmetric(horizontal: DsfrSpacings.s2w),
@@ -98,7 +98,7 @@ class _BestCarOptionView extends StatelessWidget {
   final CarSimulatorOption? bestEmissionsOption;
 
   @override
-  Widget build(final BuildContext context) => Column(
+  Widget build(final context) => Column(
     crossAxisAlignment: CrossAxisAlignment.start,
     spacing: DsfrSpacings.s2w,
     children: [
@@ -162,7 +162,7 @@ class _CarSimulatorOptionView extends StatelessWidget {
   final CarSimulatorOptionKind kind;
 
   @override
-  Widget build(final BuildContext context) {
+  Widget build(final context) {
     final currentCarIsBest =
         kind == CarSimulatorOptionKind.bestCost ? currentCar.cost <= option.cost : currentCar.emissions <= option.emissions;
 
@@ -226,7 +226,7 @@ class _KindTagView extends StatelessWidget {
   final CarSimulatorOptionKind kind;
 
   @override
-  Widget build(final BuildContext context) => switch (kind) {
+  Widget build(final context) => switch (kind) {
     CarSimulatorOptionKind.bestCost => const DsfrTag.md(
       label: TextSpan(
         text: Localisation.laPlusEconomique,
@@ -256,7 +256,7 @@ class _DiffInTag extends StatelessWidget {
   final String unit;
 
   @override
-  Widget build(final BuildContext context) {
+  Widget build(final context) {
     final diff = unit == '%' ? (to - from) / from * 100 : to - from;
     final colors =
         diff > 0
@@ -284,7 +284,7 @@ class _CurrentCarResultView extends StatelessWidget {
   final CarInfos currentCar;
 
   @override
-  Widget build(final BuildContext context) => Column(
+  Widget build(final context) => Column(
     crossAxisAlignment: CrossAxisAlignment.start,
     spacing: DsfrSpacings.s2w,
     children: [
@@ -317,7 +317,7 @@ class _NumberWithUnit extends StatelessWidget {
   final String unit;
 
   @override
-  Widget build(final BuildContext context) => Text.rich(
+  Widget build(final context) => Text.rich(
     TextSpan(
       text: FnvNumberFormat.formatNumberAfterRounding(num),
       style: const DsfrTextStyle.body2XlBold(),
@@ -332,7 +332,7 @@ class _ContextInfosView extends StatelessWidget {
   final CarInfos carInfos;
 
   @override
-  Widget build(final BuildContext context) => Container(
+  Widget build(final context) => Container(
     padding: const EdgeInsets.only(top: DsfrSpacings.s2w),
     decoration: const ShapeDecoration(shape: Border(top: BorderSide(color: DsfrColors.blueFrance950))),
     width: 270,
@@ -354,7 +354,7 @@ class _ContextInfo extends StatelessWidget {
   final String label;
 
   @override
-  Widget build(final BuildContext context) => DsfrTag.md(
+  Widget build(final context) => DsfrTag.md(
     label: TextSpan(text: label),
     backgroundColor: DsfrColors.blueFrance950,
     foregroundColor: DsfrColors.blueFranceSun113,
