@@ -32,42 +32,48 @@ class _MosaicSetState extends State<_MosaicSet> {
   @override
   void initState() {
     super.initState();
-    _responses = widget.responses;
+    _responses = List.of(widget.responses);
   }
 
   @override
-  Widget build(final context) => Wrap(
-    spacing: DsfrSpacings.s2w,
-    runSpacing: DsfrSpacings.s2w,
-    children:
-        _responses.map((final e) {
-          const size = 40.0;
+  Widget build(final context) => GridView.builder(
+    physics: const NeverScrollableScrollPhysics(),
+    shrinkWrap: true,
+    gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+      crossAxisCount: 2,
+      mainAxisSpacing: DsfrSpacings.s2w,
+      crossAxisSpacing: DsfrSpacings.s2w,
+    ),
+    itemBuilder: (final context, final index) {
+      final e = _responses[index];
+      const size = 40.0;
 
-          return MosaicButton(
-            emoji: FnvImage.network(e.imageUrl, width: size, height: size),
-            title: e.label,
-            value: e.isSelected,
-            onChanged: (final value) {
-              setState(() {
-                _responses =
-                    _responses
-                        .map(
-                          (final r) =>
-                              r.code == e.code
-                                  ? ResponseMosaic(
-                                    code: r.code,
-                                    label: r.label,
-                                    emoji: r.emoji,
-                                    imageUrl: r.imageUrl,
-                                    isSelected: value,
-                                  )
-                                  : r,
-                        )
-                        .toList();
-                context.read<QuestionEditBloc>().add(QuestionEditMosaicChangee(_responses));
-              });
-            },
-          );
-        }).toList(),
+      return MosaicButton(
+        emoji: FnvImage.network(e.imageUrl, width: size, height: size),
+        title: e.label,
+        value: e.isSelected,
+        onChanged: (final value) {
+          setState(() {
+            _responses =
+                _responses
+                    .map(
+                      (final r) =>
+                          r.code == e.code
+                              ? ResponseMosaic(
+                                code: r.code,
+                                label: r.label,
+                                emoji: r.emoji,
+                                imageUrl: r.imageUrl,
+                                isSelected: value,
+                              )
+                              : r,
+                    )
+                    .toList();
+          });
+          context.read<QuestionEditBloc>().add(QuestionEditMosaicChangee(_responses));
+        },
+      );
+    },
+    itemCount: _responses.length,
   );
 }
