@@ -13,11 +13,12 @@ final class QuizState extends Equatable {
           id: '',
           thematique: '',
           question: '',
-          reponses: [],
+          responses: [],
           points: 0,
           explicationOk: '',
           explicationKo: '',
           article: null,
+          sources: [],
         ),
         reponse: const None(),
         estExacte: const None(),
