@@ -1,6 +1,7 @@
 import 'package:app/features/theme/core/domain/mission_liste.dart';
 import 'package:app/features/theme/core/domain/service_item.dart';
 import 'package:app/features/theme/core/domain/theme_info.dart';
+import 'package:app/features/theme/core/domain/theme_summary.dart';
 import 'package:equatable/equatable.dart';
 import 'package:meta/meta.dart';
 
@@ -24,17 +25,27 @@ final class ThemeLoadInProgress extends ThemeState {
 
 @immutable
 final class ThemeLoadSuccess extends ThemeState {
-  const ThemeLoadSuccess({required this.theme, required this.missions, required this.services});
+  const ThemeLoadSuccess({required this.theme, required this.summary, required this.missions, required this.services});
 
   final ThemeInfo theme;
+  final ThemeSummary summary;
   final List<MissionListe> missions;
   final List<ServiceItem> services;
 
-  ThemeLoadSuccess copyWith({final ThemeInfo? theme, final List<MissionListe>? missions, final List<ServiceItem>? services}) =>
-      ThemeLoadSuccess(theme: theme ?? this.theme, missions: missions ?? this.missions, services: services ?? this.services);
+  ThemeLoadSuccess copyWith({
+    final ThemeInfo? theme,
+    final ThemeSummary? summary,
+    final List<MissionListe>? missions,
+    final List<ServiceItem>? services,
+  }) => ThemeLoadSuccess(
+    theme: theme ?? this.theme,
+    summary: summary ?? this.summary,
+    missions: missions ?? this.missions,
+    services: services ?? this.services,
+  );
 
   @override
-  List<Object> get props => [theme, missions, services];
+  List<Object> get props => [theme, summary, missions, services];
 }
 
 @immutable
