@@ -29,11 +29,11 @@ void main() {
           id: '42',
           thematique: '🥦 Alimentation',
           question: question,
-          reponses: [
-            QuizReponse(reponse: 'Manger moins de produits de saison', exact: false),
-            QuizReponse(reponse: 'Diminuer la consommation de viande', exact: true),
-            QuizReponse(reponse: 'Privilégier la volaille et le porc', exact: false),
-            QuizReponse(reponse: 'Réduire ses déchets', exact: false),
+          responses: [
+            QuizResponse(response: 'Manger moins de produits de saison', exact: false),
+            QuizResponse(response: 'Diminuer la consommation de viande', exact: true),
+            QuizResponse(response: 'Privilégier la volaille et le porc', exact: false),
+            QuizResponse(response: 'Réduire ses déchets', exact: false),
           ],
           points: 5,
           explicationOk:
@@ -41,6 +41,7 @@ void main() {
           explicationKo:
               "<p><span>Au contraire ! Pour rendre notre alimentation plus durable, nous pouvons manger davantage de produits de saison et augmenter la part de repas végétariens dans les menus de la semaine. Diminuer notre consommation de viande permet en effet de réduire les impacts écologiques du secteur de l’élevage, qui génère à lui seul près de 15 % des émissions mondiales de gaz à effet de serre : c'est donc l'action la plus efficace pour limiter l'impact de notre alimentation.<br><br>On peut aussi privilégier les produits locaux et biologiques, limiter l’achat de produits transformés et réduire le gaspillage alimentaire.</span></p>",
           article: null,
+          sources: [],
         ),
       );
       await _init(tester, titre);
@@ -58,11 +59,11 @@ void main() {
         id: '42',
         thematique: '🥦 Alimentation',
         question: 'Quelle action est la plus efficace pour une alimentation plus durable ?',
-        reponses: [
-          QuizReponse(reponse: 'Manger moins de produits de saison', exact: false),
-          QuizReponse(reponse: reponse, exact: true),
-          QuizReponse(reponse: 'Privilégier la volaille et le porc', exact: false),
-          QuizReponse(reponse: 'Réduire ses déchets', exact: false),
+        responses: [
+          QuizResponse(response: 'Manger moins de produits de saison', exact: false),
+          QuizResponse(response: reponse, exact: true),
+          QuizResponse(response: 'Privilégier la volaille et le porc', exact: false),
+          QuizResponse(response: 'Réduire ses déchets', exact: false),
         ],
         points: 5,
         explicationOk:
@@ -71,6 +72,7 @@ void main() {
             "<p><span>Au contraire ! Pour rendre notre alimentation plus durable, nous pouvons manger davantage de produits de saison et augmenter la part de repas végétariens dans les menus de la semaine. Diminuer notre consommation de viande permet en effet de réduire les impacts écologiques du secteur de l’élevage, qui génère à lui seul près de 15 % des émissions mondiales de gaz à effet de serre : c'est donc l'action la plus efficace pour limiter l'impact de notre alimentation.<br><br>On peut aussi privilégier les produits locaux et biologiques, limiter l’achat de produits transformés et réduire le gaspillage alimentaire.</span></p>",
         article:
             "<p>Que l’on se rende dans un magasin de quartier pour faire ses courses, qu’on allume la lumière ou qu’on chauffe son logement, l’ensemble de nos actions quotidiennes a un impact sur l’environnement. Cet impact, c’est ce que l’on appelle l’<strong>empreinte carbone</strong>, comme une trace invisible que nous laissons derrière nous.</p><p>Elle mesure la quantité totale de gaz à effet de serre (dioxyde de carbone (CO2), protoxyde d'azote, méthane, ...) liés à notre consommation et permet donc de quantifier nos émissions selon notre mode de vie. Elle peut concerner les émissions d’un individu (son mode de vie), d’une entreprise (ses activités) ou d’une population, d'un territoire.</p><ul><li><p>Au niveau des entreprises, la loi Grenelle II impose depuis juillet 2020 le Bilan GES Réglementaire à un nombre de structures publiques et privées. Cela concerne les entreprises publiques de plus de 250 personnes, 500 personnes pour les privées (250 en outre-mer), les collectivités de plus de 50 000 habitants et l’État.</p></li><li><p>A l'échelle du pays, l'empreinte carbone moyenne d'un Français est estimée à <strong>8,9 tonnes d'équivalent CO2</strong> en 2021. Or, pour respecter les objectifs de l’Accord de Paris et maintenir le réchauffement planétaire sous les 2°C, il nous faudrait réduire ce nombre à <strong>deux tonnes</strong>, autrement dit : le diviser presque par cinq !</p></li></ul><p><em>Pour en savoir plus sur l'empreinte carbone de la France, </em><a target=\"_self\" rel=\"\" href=\"/article/L'empreinte carbone de la France/11\"><em>c'est par ici.</em></a></p><h2>Comment connaître votre empreinte carbone ?</h2><p>L’empreinte carbone se calcule aussi au niveau individuel, et elle dépend directement d’un ensemble d’activités que nous effectuons quotidiennement. Calculer son empreinte carbone, c'est prendre conscience de ses activités quotidiennes et de leurs conséquences pour l'environnement.</p><p>On peut ainsi cibler les activités les plus polluantes et ajuster nos usages et nos habitudes de consommation pour diminuer notre impact. Faites le bilan avec le calculateur de l'Ademe : <a target=\"_blank\" rel=\"\" class=\"in-cell-link\" href=\"https://nosgestesclimat.fr/\"><span style=\"color: rgb(17, 85, 204)\">Nos Gestes Climat</span></a>.</p>",
+        sources: [],
       );
       ielALeQuizSuivant(quiz);
 
@@ -80,12 +82,7 @@ void main() {
 
       ielVoitLeTexte(Localisation.pourquoi);
 
-      verify(
-        () => ScenarioContext().dioMock!.post<dynamic>(
-          Endpoints.events,
-          data: '{"content_id":"42","number_value":100,"type":"quizz_score"}',
-        ),
-      );
+      verify(() => ScenarioContext().dioMock!.patch<dynamic>(Endpoints.quiz(quiz.id), data: {'pourcent': 100}));
     });
   });
 
@@ -95,36 +92,31 @@ void main() {
       const titre = 'Une assiette plus durable';
       ielALesRecommandationsSuivantes(titre);
       const reponse = 'Diminuer la consommation de viande';
-      ielALeQuizSuivant(
-        const Quiz(
-          id: '42',
-          thematique: '🥦 Alimentation',
-          question: 'Quelle action est la plus efficace pour une alimentation plus durable ?',
-          reponses: [
-            QuizReponse(reponse: 'Manger moins de produits de saison', exact: false),
-            QuizReponse(reponse: reponse, exact: true),
-            QuizReponse(reponse: 'Privilégier la volaille et le porc', exact: false),
-            QuizReponse(reponse: 'Réduire ses déchets', exact: false),
-          ],
-          points: 5,
-          explicationOk:
-              '<p><span>Le secteur de l’élevage génère à lui seul près de 15 % des émissions mondiales de gaz à effet de serre. Réduire notre consommation de viande permet d’agir sur la production et de diminuer les impacts qui lui sont associés.<br><br>Pour rendre notre alimentation plus durable, on peut aussi privilégier les produits locaux, de saison et biologiques, limiter l’achat de produits transformés et réduire le gaspillage alimentaire.</span></p>',
-          explicationKo:
-              "<p><span>Au contraire ! Pour rendre notre alimentation plus durable, nous pouvons manger davantage de produits de saison et augmenter la part de repas végétariens dans les menus de la semaine. Diminuer notre consommation de viande permet en effet de réduire les impacts écologiques du secteur de l’élevage, qui génère à lui seul près de 15 % des émissions mondiales de gaz à effet de serre : c'est donc l'action la plus efficace pour limiter l'impact de notre alimentation.<br><br>On peut aussi privilégier les produits locaux et biologiques, limiter l’achat de produits transformés et réduire le gaspillage alimentaire.</span></p>",
-          article: null,
-        ),
+      const quiz = Quiz(
+        id: '42',
+        thematique: '🥦 Alimentation',
+        question: 'Quelle action est la plus efficace pour une alimentation plus durable ?',
+        responses: [
+          QuizResponse(response: 'Manger moins de produits de saison', exact: false),
+          QuizResponse(response: reponse, exact: true),
+          QuizResponse(response: 'Privilégier la volaille et le porc', exact: false),
+          QuizResponse(response: 'Réduire ses déchets', exact: false),
+        ],
+        points: 5,
+        explicationOk:
+            '<p><span>Le secteur de l’élevage génère à lui seul près de 15 % des émissions mondiales de gaz à effet de serre. Réduire notre consommation de viande permet d’agir sur la production et de diminuer les impacts qui lui sont associés.<br><br>Pour rendre notre alimentation plus durable, on peut aussi privilégier les produits locaux, de saison et biologiques, limiter l’achat de produits transformés et réduire le gaspillage alimentaire.</span></p>',
+        explicationKo:
+            "<p><span>Au contraire ! Pour rendre notre alimentation plus durable, nous pouvons manger davantage de produits de saison et augmenter la part de repas végétariens dans les menus de la semaine. Diminuer notre consommation de viande permet en effet de réduire les impacts écologiques du secteur de l’élevage, qui génère à lui seul près de 15 % des émissions mondiales de gaz à effet de serre : c'est donc l'action la plus efficace pour limiter l'impact de notre alimentation.<br><br>On peut aussi privilégier les produits locaux et biologiques, limiter l’achat de produits transformés et réduire le gaspillage alimentaire.</span></p>",
+        article: null,
+        sources: [],
       );
+      ielALeQuizSuivant(quiz);
       await _init(tester, titre);
       await ielAppuieSur(tester, reponse);
       await ielAppuieSur(tester, Localisation.valider);
       ielVoitLeTexte(Localisation.pourquoi);
 
-      verify(
-        () => ScenarioContext().dioMock!.post<dynamic>(
-          Endpoints.events,
-          data: '{"content_id":"42","number_value":100,"type":"quizz_score"}',
-        ),
-      );
+      verify(() => ScenarioContext().dioMock!.patch<dynamic>(Endpoints.quiz(quiz.id), data: {'pourcent': 100}));
     });
   });
 
@@ -134,44 +126,38 @@ void main() {
       const titre = 'Une assiette plus durable';
       ielALesRecommandationsSuivantes(titre);
       const reponse = 'Manger moins de produits de saison';
-      ielALeQuizSuivant(
-        const Quiz(
-          id: '42',
-          thematique: '🥦 Alimentation',
-          question: 'Quelle action est la plus efficace pour une alimentation plus durable ?',
-          reponses: [
-            QuizReponse(reponse: reponse, exact: false),
-            QuizReponse(reponse: 'Diminuer la consommation de viande', exact: true),
-            QuizReponse(reponse: 'Privilégier la volaille et le porc', exact: false),
-            QuizReponse(reponse: 'Réduire ses déchets', exact: false),
-          ],
-          points: 5,
-          explicationOk:
-              '<p><span>Le secteur de l’élevage génère à lui seul près de 15 % des émissions mondiales de gaz à effet de serre. Réduire notre consommation de viande permet d’agir sur la production et de diminuer les impacts qui lui sont associés.<br><br>Pour rendre notre alimentation plus durable, on peut aussi privilégier les produits locaux, de saison et biologiques, limiter l’achat de produits transformés et réduire le gaspillage alimentaire.</span></p>',
-          explicationKo:
-              "<p><span>Au contraire ! Pour rendre notre alimentation plus durable, nous pouvons manger davantage de produits de saison et augmenter la part de repas végétariens dans les menus de la semaine. Diminuer notre consommation de viande permet en effet de réduire les impacts écologiques du secteur de l’élevage, qui génère à lui seul près de 15 % des émissions mondiales de gaz à effet de serre : c'est donc l'action la plus efficace pour limiter l'impact de notre alimentation.<br><br>On peut aussi privilégier les produits locaux et biologiques, limiter l’achat de produits transformés et réduire le gaspillage alimentaire.</span></p>",
-          article:
-              "<p>Que l’on se rende dans un magasin de quartier pour faire ses courses, qu’on allume la lumière ou qu’on chauffe son logement, l’ensemble de nos actions quotidiennes a un impact sur l’environnement. Cet impact, c’est ce que l’on appelle l’<strong>empreinte carbone</strong>, comme une trace invisible que nous laissons derrière nous.</p><p>Elle mesure la quantité totale de gaz à effet de serre (dioxyde de carbone (CO2), protoxyde d'azote, méthane, ...) liés à notre consommation et permet donc de quantifier nos émissions selon notre mode de vie. Elle peut concerner les émissions d’un individu (son mode de vie), d’une entreprise (ses activités) ou d’une population, d'un territoire.</p><ul><li><p>Au niveau des entreprises, la loi Grenelle II impose depuis juillet 2020 le Bilan GES Réglementaire à un nombre de structures publiques et privées. Cela concerne les entreprises publiques de plus de 250 personnes, 500 personnes pour les privées (250 en outre-mer), les collectivités de plus de 50 000 habitants et l’État.</p></li><li><p>A l'échelle du pays, l'empreinte carbone moyenne d'un Français est estimée à <strong>8,9 tonnes d'équivalent CO2</strong> en 2021. Or, pour respecter les objectifs de l’Accord de Paris et maintenir le réchauffement planétaire sous les 2°C, il nous faudrait réduire ce nombre à <strong>deux tonnes</strong>, autrement dit : le diviser presque par cinq !</p></li></ul><p><em>Pour en savoir plus sur l'empreinte carbone de la France, </em><a target=\"_self\" rel=\"\" href=\"/article/L'empreinte carbone de la France/11\"><em>c'est par ici.</em></a></p><h2>Comment connaître votre empreinte carbone ?</h2><p>L’empreinte carbone se calcule aussi au niveau individuel, et elle dépend directement d’un ensemble d’activités que nous effectuons quotidiennement. Calculer son empreinte carbone, c'est prendre conscience de ses activités quotidiennes et de leurs conséquences pour l'environnement.</p><p>On peut ainsi cibler les activités les plus polluantes et ajuster nos usages et nos habitudes de consommation pour diminuer notre impact. Faites le bilan avec le calculateur de l'Ademe : <a target=\"_blank\" rel=\"\" class=\"in-cell-link\" href=\"https://nosgestesclimat.fr/\"><span style=\"color: rgb(17, 85, 204)\">Nos Gestes Climat</span></a>.</p>",
-        ),
+      const quiz = Quiz(
+        id: '42',
+        thematique: '🥦 Alimentation',
+        question: 'Quelle action est la plus efficace pour une alimentation plus durable ?',
+        responses: [
+          QuizResponse(response: reponse, exact: false),
+          QuizResponse(response: 'Diminuer la consommation de viande', exact: true),
+          QuizResponse(response: 'Privilégier la volaille et le porc', exact: false),
+          QuizResponse(response: 'Réduire ses déchets', exact: false),
+        ],
+        points: 5,
+        explicationOk:
+            '<p><span>Le secteur de l’élevage génère à lui seul près de 15 % des émissions mondiales de gaz à effet de serre. Réduire notre consommation de viande permet d’agir sur la production et de diminuer les impacts qui lui sont associés.<br><br>Pour rendre notre alimentation plus durable, on peut aussi privilégier les produits locaux, de saison et biologiques, limiter l’achat de produits transformés et réduire le gaspillage alimentaire.</span></p>',
+        explicationKo:
+            "<p><span>Au contraire ! Pour rendre notre alimentation plus durable, nous pouvons manger davantage de produits de saison et augmenter la part de repas végétariens dans les menus de la semaine. Diminuer notre consommation de viande permet en effet de réduire les impacts écologiques du secteur de l’élevage, qui génère à lui seul près de 15 % des émissions mondiales de gaz à effet de serre : c'est donc l'action la plus efficace pour limiter l'impact de notre alimentation.<br><br>On peut aussi privilégier les produits locaux et biologiques, limiter l’achat de produits transformés et réduire le gaspillage alimentaire.</span></p>",
+        article:
+            "<p>Que l’on se rende dans un magasin de quartier pour faire ses courses, qu’on allume la lumière ou qu’on chauffe son logement, l’ensemble de nos actions quotidiennes a un impact sur l’environnement. Cet impact, c’est ce que l’on appelle l’<strong>empreinte carbone</strong>, comme une trace invisible que nous laissons derrière nous.</p><p>Elle mesure la quantité totale de gaz à effet de serre (dioxyde de carbone (CO2), protoxyde d'azote, méthane, ...) liés à notre consommation et permet donc de quantifier nos émissions selon notre mode de vie. Elle peut concerner les émissions d’un individu (son mode de vie), d’une entreprise (ses activités) ou d’une population, d'un territoire.</p><ul><li><p>Au niveau des entreprises, la loi Grenelle II impose depuis juillet 2020 le Bilan GES Réglementaire à un nombre de structures publiques et privées. Cela concerne les entreprises publiques de plus de 250 personnes, 500 personnes pour les privées (250 en outre-mer), les collectivités de plus de 50 000 habitants et l’État.</p></li><li><p>A l'échelle du pays, l'empreinte carbone moyenne d'un Français est estimée à <strong>8,9 tonnes d'équivalent CO2</strong> en 2021. Or, pour respecter les objectifs de l’Accord de Paris et maintenir le réchauffement planétaire sous les 2°C, il nous faudrait réduire ce nombre à <strong>deux tonnes</strong>, autrement dit : le diviser presque par cinq !</p></li></ul><p><em>Pour en savoir plus sur l'empreinte carbone de la France, </em><a target=\"_self\" rel=\"\" href=\"/article/L'empreinte carbone de la France/11\"><em>c'est par ici.</em></a></p><h2>Comment connaître votre empreinte carbone ?</h2><p>L’empreinte carbone se calcule aussi au niveau individuel, et elle dépend directement d’un ensemble d’activités que nous effectuons quotidiennement. Calculer son empreinte carbone, c'est prendre conscience de ses activités quotidiennes et de leurs conséquences pour l'environnement.</p><p>On peut ainsi cibler les activités les plus polluantes et ajuster nos usages et nos habitudes de consommation pour diminuer notre impact. Faites le bilan avec le calculateur de l'Ademe : <a target=\"_blank\" rel=\"\" class=\"in-cell-link\" href=\"https://nosgestesclimat.fr/\"><span style=\"color: rgb(17, 85, 204)\">Nos Gestes Climat</span></a>.</p>",
+        sources: [],
       );
+      ielALeQuizSuivant(quiz);
       await _init(tester, titre);
       await ielAppuieSur(tester, reponse);
       await ielAppuieSur(tester, Localisation.valider);
       ielVoitLeTexte(Localisation.pourquoi);
-
-      verify(
-        () => ScenarioContext().dioMock!.post<dynamic>(
-          Endpoints.events,
-          data: '{"content_id":"42","number_value":0,"type":"quizz_score"}',
-        ),
-      );
+      verify(() => ScenarioContext().dioMock!.patch<dynamic>(Endpoints.quiz(quiz.id), data: {'pourcent': 0}));
     });
   });
 }
 
 Future<void> _init(final WidgetTester tester, final String titre) async {
   ielEstConnecte();
-  ScenarioContext().dioMock!.postM(Endpoints.events);
+  ScenarioContext().dioMock!.patchM(Endpoints.quiz('42'));
   await ielLanceLapplication(tester);
   await iTapOn(tester, '🍛 Me nourrir');
   await ielScrolle(tester, titre);
