@@ -14,7 +14,7 @@ void ielALeQuizSuivant(final Quiz valeur) {
       'questions': [
         {
           'libelle': valeur.question,
-          'reponses': valeur.reponses.map((final e) => {'reponse': e.reponse, 'exact': e.exact}).toList(),
+          'reponses': valeur.responses.map((final e) => {'reponse': e.response, 'exact': e.exact}).toList(),
           'explicationOk': valeur.explicationOk,
           'explicationKO': valeur.explicationKo,
         },
