@@ -12,10 +12,10 @@ class DsfrStepper extends StatelessWidget {
   final int total;
 
   @override
-  Widget build(final BuildContext context) => Column(
+  Widget build(final context) => Column(
     crossAxisAlignment: CrossAxisAlignment.start,
     children: [
-      Text('Étape $current sur $total', style: const DsfrTextStyle.bodySm(color: DsfrColors.grey200)),
+      Text('Question $current sur $total', style: const DsfrTextStyle.bodySm(color: DsfrColors.grey200)),
       const SizedBox(height: DsfrSpacings.s1v),
       Text(title, style: const DsfrTextStyle.headline6()),
       const SizedBox(height: DsfrSpacings.s3v),
