@@ -413,14 +413,14 @@ class DsfrColors {
   static const backgroundContrastYellowTournesol = Color(0xFFFEECC2);
   static const textActionHighYellowMoutarde = Color(0xFF695240);
   static const backgroundContrastYellowMoutarde = Color(0xFFFEEBD0);
-  static const textActionHighGreenBourgeon = Color(0xFF447049);
-  static const backgroundContrastGreenBourgeon = Color(0xFFC9FCAC);
-  static const textDefaultSuccess = Color(0xFF18753C);
-  static const backgroundConstrastSuccess = Color(0xFFB8FEC9);
-  static const textDefaultError = Color(0xFFCE0500);
-  static const backgroundConstrastError = Color(0xFFFFE9E9);
-  static const borderPlainInfo = Color(0xFF0063CB);
-  static const backgroundFlatInfo = Color(0xFF0063CB);
   static const backgroundContrastGrey = Color(0xFFEEEEEE);
   static const textDefaultGrey = Color(0xFF3A3A3A);
+  static const textActionHighGreenBourgeon = greenBourgeonSun425;
+  static const backgroundContrastGreenBourgeon = greenBourgeon950;
+  static const textDefaultSuccess = success425;
+  static const backgroundConstrastSuccess = success950;
+  static const textDefaultError = error425;
+  static const backgroundConstrastError = Color(0xFFFFE9E9);
+  static const borderPlainInfo = info425;
+  static const backgroundFlatInfo = info425;
 }
